+++ src/app.js
@@ -1,10 +1,10 @@
 import Store from './store.js';
-import TodoAction from './todoAction.js';
-import TodoRender from './todoRender.js';
+import TodoReducer from './todoReducer.js';
+import TodoViewer from './todoViewer.js';
 
 (function () {
     const store = new Store();
 
-    new TodoAction(store);
-    new TodoRender(store);
+    new TodoReducer(store);
+    new TodoViewer(store);
 })();
