+++ src/components/TodoList.js
@@ -1,39 +1,17 @@
-import { getElement } from './util.js';
-import { FILTER_TYPE, UI_CLASS, KEY_CODE, MESSAGES } from './constant.js';
+import { getElement } from '../util.js';
+import { UI_CLASS, KEY_CODE, MESSAGES } from '../constant.js';
 
-class TodoAction {
+class TodoList {
     constructor(store) {
         this.store = store;
-        this.inputEl = getElement('input.new-todo');
-        this.filtersEl = getElement('ul.filters');
         this.todoListEl = getElement('ul.todo-list');
         this.init();
     }
 
     init() {
-        this.inputEl.addEventListener('keyup', this.addTodoHandler.bind(this));
         this.todoListEl.addEventListener('click', this.todoClickDelegationHandler.bind(this));
         this.todoListEl.addEventListener('dblclick', this.modifyHandler.bind(this));
         this.todoListEl.addEventListener('keyup', this.confirmHandler.bind(this));
-        this.filtersEl.addEventListener('click', this.filtersHandler.bind(this));
-    }
-
-    addTodoHandler({ keyCode, target }) {
-        if (keyCode !== KEY_CODE.ENTER || !target.value) return;
-        const todoList = this.store.get().todoList;
-        const id = new Date().getTime();
-        const newTodo = {
-            title: target.value,
-            id: id,
-            isCompleted: false,
-            isEditing: false
-        };
-        todoList[id] = newTodo;
-        target.value = '';
-
-        this.store.set({
-            todoList: { ...todoList }
-        });
     }
 
     _toggleTodoItem({ id }) {
@@ -83,20 +61,6 @@ class TodoAction {
             });
         }
     }
-
-    filtersHandler({ target }) {
-        if (target.tagName !== 'A') return;
-        const type = target.classList[0];
-
-        this.filtersEl.querySelectorAll('a').forEach(el => {
-            el.classList.remove(UI_CLASS.SELECTED);
-            if (el.classList.contains(type)) el.classList.add(UI_CLASS.SELECTED);
-        });
-
-        this.store.set({
-            filter: FILTER_TYPE[type.toUpperCase()]
-        });
-    }
 }
 
-export default TodoAction
+export default TodoList
